requirements.txt
fastapi
uvicorn 
python-multipart 
python-docx
Spire.Doc
pdf2docx